% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSeqfromSp2.R
\name{simSeqfromSp2}
\alias{simSeqfromSp2}
\title{simSeqfromSp2: Modified function of simSeqfromSp from R package "phybase". Please cite:XXX}
\usage{
simSeqfromSp2(sptree, spname, ntaxasp, ngene, theta = 0, noclock = 0,
  simsequence = 1, murate = "Dirichlet", alpha = 5, seqlength = 100,
  model = 1, kappa = 2, rate = c(1, 1, 1, 1, 1, 1), frequency = c(1/4,
  1/4, 1/4, 1/4), outfile, format = "phylip")
}
\arguments{
\item{sptree}{A species tree which must be a rooted tree.}

\item{spname}{species names}

\item{ntaxasp}{a vector of the number of individuals in each species}

\item{ngene}{number of genes}

\item{theta}{population size}

\item{noclock}{0: clocklike species tree 1: nonclocklike species tree}

\item{simsequence}{1: simulate sequences and gene trees, 0: simulate gene trees}

\item{murate}{distribution of mutation rates}

\item{alpha}{the shape parameter of dirichlet distribution}

\item{seqlength}{the number of nucleotides along the sequences}

\item{model}{substitution model}

\item{kappa}{transition/transversion ratio}

\item{rate}{rates}

\item{frequency}{nucleotide frequency}

\item{outfile}{the full path of the output file}

\item{format}{either "phylip" or "nexus"}
}
\description{
The function simulates sequences from a species tree.
}
\keyword{phybase}

